+++ android/src/main/java/flutter/overlay/window/flutter_overlay_window/WindowSetup.java
@@ -6,15 +6,12 @@ import android.view.WindowManager;
 
 import androidx.core.app.NotificationCompat;
 
-import io.flutter.plugin.common.BasicMessageChannel;
-
 public abstract class WindowSetup {
 
     static int height = WindowManager.LayoutParams.MATCH_PARENT;
     static int width = WindowManager.LayoutParams.MATCH_PARENT;
     static int flag = WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE;
     static int gravity = Gravity.CENTER;
-    static BasicMessageChannel<Object> messenger = null;
     static String overlayTitle = "Overlay is activated";
     static String overlayContent = "Tap to edit settings or disable";
     static String positionGravity = "none";